fix(profile): creating third hook for onboarding state
